+++ client/src/components/Landing/Sign.js
@@ -1,0 +1,5 @@
+const Sign = () => {
+    return <></>
+}
+
+export default Sign;
